+++ src/websocket_manager.rs
@@ -16,7 +16,7 @@ use crate::db_connection_manager::establish_connection;
 // Define a structure for messages
 #[derive(Debug)]
 pub struct MyMessage {
-    pub timestamp: u128,
+    pub timestamp: u64,
     pub endpoint_name: String,
     pub message: Message,
 }
@@ -44,8 +44,8 @@ pub(crate) struct SubscribeMessage {
 
 
 pub(crate) async fn forward_general_message(text: String, uri: &str, general_tx: &mpsc::Sender<MyMessage>) {
-    let timestamp: u128 = match chrono::Utc::now().timestamp_nanos_opt() {
-        Some(nanos) => nanos as u128,
+    let timestamp: u64 = match chrono::Utc::now().timestamp_nanos_opt() {
+        Some(nanos) => nanos as u64,
         None => {
             // Handle the error appropriately, such as logging or panicking
             panic!("Unable to obtain timestamp in nanoseconds.");
